tests: extract network fixtures from tests/conftest.py
Move network fixtures into domain-specific modules under
tests/fixtures/network/:
- node_nics.py: NIC discovery (node_physical_nics, nodes_active_nics,
  nodes_available_nics, worker_nodes_ipv4_false_secondary_nics,
  hosts_common_available_ports)
- mac_pool.py: KubeMacPool (mac_pool, kmp_vm_label, kmp_enabled_ns,
  kmp_enabled_namespace, kmp_deployment)
- nmstate.py: NMState placeholder (nmstate_dependent_placeholder)
- sriov.py: SR-IOV (sriov_namespace, sriov_workers, sriov_node_policy)

Register new modules in pytest_plugins (root conftest.py).
Remove migrated fixtures and now-unused imports from tests/conftest.py.
